+++ tests/examples/detach_running_step/main.sh
@@ -2,8 +2,7 @@
 source ../example.rc
 
 # Run the example
-echo "Hello world!" > inp.txt
-sb -j 2 -w & # > current_stdout.txt &
+sb -j 3 -w & # > current_stdout.txt &
 
 # Get the graph after completion of the pending steps.
 stepup wait
@@ -15,5 +14,7 @@ wait
 
 # Check files that are expected to be present and/or missing.
 [[ -f plan.py ]] || exit 1
-[[ -f inp.txt ]] || exit 1
-[[ -f out.txt ]] || exit 1
+[[ -f provided.txt ]] || exit 1
+[[ -f trigger_work1.txt ]] || exit 1
+[[ -f trigger_sub.txt ]] || exit 1
+[[ -f trigger_work2.txt ]] || exit 1