+++ tests/examples/error_detach_running/main.sh
@@ -2,8 +2,8 @@
 source ../example.rc
 
 # Run the example
-echo "Hello world!" > inp.txt
 sb -j 2 -w & # > current_stdout.txt &
+PID=$!
 
 # Get the graph after completion of the pending steps.
 stepup wait
@@ -11,9 +11,10 @@ stepup graph current_graph
 stepup join
 
 # Wait for background processes, if any.
-wait
+set +e; wait -fn $PID; RETURNCODE=$?; set -e
+[[ "${RETURNCODE}" -eq 2 ]] || exit 1
 
 # Check files that are expected to be present and/or missing.
 [[ -f plan.py ]] || exit 1
-[[ -f inp.txt ]] || exit 1
-[[ -f out.txt ]] || exit 1
+[[ -f work.py ]] || exit 1
+[[ -f trigger_work.txt ]] || exit 1